start-crawler/requirements.txt
humanize==4.10.0
PyMySQL==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.27
